% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BitterCAS-package}
\alias{BitterCAS}
\alias{BitterCAS-package}
\title{bitterCAS}
\description{
Provides an interface to Maxima, a computer algebra system (CAS).
}
\details{
Note: You need to install the Maxima software separately in order to make use of this package.
}
\author{
\strong{Maintainer}: Manuel Ordovas \email{ordovasmanuel@gmail.com}

}
