% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineCAS.R
\name{EngineCAS}
\alias{EngineCAS}
\title{Internal only function, returns last command as latex}
\usage{
EngineCAS(options = NULL)
}
\arguments{
\item{options}{valid connection to maxima.}
}
\description{
Internal only function, returns last command as latex
}
