% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMaximaPort.R
\name{FindMaximaPort}
\alias{FindMaximaPort}
\title{Internal only function, find an available port to connect with maxima}
\usage{
FindMaximaPort(fromPort = 1024, toPort = 49151)
}
\arguments{
\item{fromPort}{integer, defaults at 1024 First port from which to look for available ports.}

\item{toPort}{integer, defaults at 49151 End port from 'fromPort' from which to look for available ports.}
}
\description{
Ports 1024-49151 are the User Ports and are the ones to use for your custom protocols.
}
