% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTexOutput.R
\name{GetTexOutput}
\alias{GetTexOutput}
\title{Internal only function, extracts tex data}
\usage{
GetTexOutput(mlines, con, waitTime = 0.2)
}
\arguments{
\item{mlines}{lines to be sent to maxima}

\item{con}{successful socket connection to maxima}

\item{waitTime}{time to wait from sending information to maxima and retreiving it}
}
\description{
Internal only function, extracts tex data
}
