% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAS.R
\name{ICAS}
\alias{ICAS}
\title{Inline R function with maxima code, wraps EngineCAS}
\usage{
ICAS(code = NULL)
}
\arguments{
\item{code}{string with maxima code.}
}
\description{
Inline R function with maxima code, wraps EngineCAS
}
