% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitCAS.R
\name{InitCAS}
\alias{InitCAS}
\title{Internal only function, prepares maxima environment}
\usage{
InitCAS(con, userDir)
}
\arguments{
\item{con}{successful socket connection to maxima}

\item{userDir}{user directory to save intermediate objects}
}
\description{
Internal only function, prepares maxima environment
}
