% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximaConnection.R
\name{MaximaConnection}
\alias{MaximaConnection}
\title{Internal only function, creates a new Socket for a connection with maxima.}
\usage{
MaximaConnection(con, port)
}
\arguments{
\item{con}{successful socket connection from FindMaximaPort()}

\item{port}{port of the successful socket connection()}
}
\description{
Internal only function, creates a new Socket for a connection with maxima.
}
