% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngineCAS.R
\name{EngineCAS}
\alias{EngineCAS}
\title{CAS engine, ran by CAS code chunks.}
\usage{
EngineCAS(options = NULL)
}
\arguments{
\item{options}{chunk options}
}
\description{
This function is only meant to be used in RMD files within chunks.
}
